Optimize for non-int and non-dup-sort types of databases using features
