+++ Cargo.toml
@@ -14,8 +14,10 @@ categories = ["database"]
 exclude = ["/tests/envs/*"]
 
 [features]
-default = []
+default = ["db-dup-sort", "db-int-key"]
 backtrace = ["failure/backtrace", "failure/std"]
+db-dup-sort = []
+db-int-key = []
 with-asan = ["lmdb-rkv/with-asan"]
 with-fuzzer = ["lmdb-rkv/with-fuzzer"]
 with-fuzzer-no-link = ["lmdb-rkv/with-fuzzer-no-link"]
