+++ src/backend/common.rs
@@ -25,7 +25,9 @@ pub enum EnvironmentFlags {
 
 pub enum DatabaseFlags {
     REVERSE_KEY,
+    #[cfg(feature = "db-dup-sort")]
     DUP_SORT,
+    #[cfg(feature = "db-int-key")]
     INTEGER_KEY,
     DUP_FIXED,
     INTEGER_DUP,
