+++ src/backend/impl_lmdb/flags.rs
@@ -84,7 +84,9 @@ impl Into<lmdb::DatabaseFlags> for DatabaseFlags {
     fn into(self) -> lmdb::DatabaseFlags {
         match self {
             DatabaseFlags::REVERSE_KEY => lmdb::DatabaseFlags::REVERSE_KEY,
+            #[cfg(feature = "db-dup-sort")]
             DatabaseFlags::DUP_SORT => lmdb::DatabaseFlags::DUP_SORT,
+            #[cfg(feature = "db-int-key")]
             DatabaseFlags::INTEGER_KEY => lmdb::DatabaseFlags::INTEGER_KEY,
             DatabaseFlags::DUP_FIXED => lmdb::DatabaseFlags::DUP_FIXED,
             DatabaseFlags::INTEGER_DUP => lmdb::DatabaseFlags::INTEGER_DUP,
